+++ unix/Xvnc/programs/Xserver/include/timetrack.h
@@ -13,7 +13,7 @@ typedef char *XPointer;
 typedef struct timeTrack{
     int eventID;
     int valid;
-    unsigned long array[TIME_COLUM];
+    unsigned int array[TIME_COLUM];
 }timeTrack;
